完善编码器功能

+++ WARSHIP/Encoder/Encoder.h
@@ -3,7 +3,7 @@
 #include "sys.h"
 
 void Orth_Encoder_Init(TIM_TypeDef *TIMx);
-u16 Orth_Get_EnCount(TIM_TypeDef *TIMx);
-void Orth_Set_EnCount(TIM_TypeDef *TIMx,u16 count);
+int Orth_Get_EnCount(TIM_TypeDef *TIMx);
+void Orth_Set_EnCount(TIM_TypeDef *TIMx,int count);
 
 #endif
